+++ src/Components/TrainingInstanceBuilder.jsx
@@ -1,30 +1,14 @@
 import { useContext, useEffect, useState } from "react";
-import { getTrainings } from "../adapter";
-import { TokenContext } from "./TokenContext";
+import TrainingsScroller from "./TrainingsScroller";
 
 export default function ({ build }) {
-  const token = useContext(TokenContext);
-  const [trainings, setTrainings] = useState([]);
   const [cursor, setCursor] = useState(undefined);
 
-  const fn = async () => {
-    const trainings = await getTrainings({
-      token,
-      offset: 0,
-      limit: 20,
-      order_by: "name",
-      descending: true,
-    });
-    setTrainings(trainings);
-  };
-  useEffect(() => {
-    fn();
-  }, []);
+  const LIMIT_SIZE = 8;
 
-  return (
-    <>
-      <div className="font-bold">すべてのトレーニング</div>
-      {trainings.map((training) => (
+  const createDef = (trainings) => {
+    return(
+      trainings.map((training) => (
         <div
           onClick={(_) => setCursor(training)}
           className="mb-3 rounded-2xl bg-gradient-to-br from-bright_accent to-accent p-3"
@@ -33,7 +17,15 @@ export default function ({ build }) {
           <div>重量(デフォルト): {training.weight}</div>
           <div>回数(デフォルト): {training.times}</div>
         </div>
-      ))}
+      ))
+    );
+  }
+
+  return (
+    <>
+      <div className="font-bold">すべてのトレーニング</div>
+      
+      <TrainingsScroller limit={LIMIT_SIZE} def={createDef}/>
 
       {cursor ? (
         <>
